requirements.txt
aws-cdk-lib==2.14.0
aws-cdk.aws-s3
aws-cdk.aws-ssm
aws-cdk.aws-ec2
aws-cdk.aws-iam
aws-cdk.aws-kms
constructs>=10.0.0,<11.0.0
